# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Task Assistant Worker API
  description: |
    Cloudflare Workers + Hono 后端 HTTP API（PRD v1.2 / 技术方案 v1.4 对齐）。

    **鉴权**：多数 `/api/*` 需 `Authorization: Bearer <token>`。`token` 为 **JWT**（`POST /api/auth/login` 签发，`sub` = `users.id`）或开发用 **明文用户 id**（如 `local-dev-user`，须存在于 D1）。

    **SSE 对话**：`POST /api/chat/stream` 返回 `text/event-stream`；帧格式为 `event: <名称>` + `data: <JSON>` + 空行。事件类型见下方该路径说明及 `GET /health/sse-chat`。

    **管理 Prompt**：`/api/prompts/*` 需配置 Worker 环境变量 `ADMIN_API_SECRET`，并使用 `Authorization: Bearer <ADMIN_API_SECRET>` 或 `X-Admin-Token: <ADMIN_API_SECRET>`。
  version: 1.2.0
servers:
  - url: http://127.0.0.1:8787
    description: 本地 wrangler dev
  - url: https://{workerSubdomain}.workers.dev
    description: Cloudflare Workers 默认子域（部署后替换）
    variables:
      workerSubdomain:
        default: your-worker
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT（login）或明文 users.id（仅开发）
...
paths:
...
  /api/workspace:
    get:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 工作空间文件列表
      description: |
        `folder` 与 `folder_path` 等价；`folder` 为空串表示仅根目录；非空为 `folder_path` 前缀匹配。
      parameters:
        - name: folder
          in: query
          schema: { type: string }
        - name: folder_path
          in: query
          schema: { type: string }
        - name: type
          in: query
          description: semantic_type 过滤
          schema: { type: string }
      responses:
        '200':
          description: 文件元数据数组（含 processed 等）
  /api/workspace/events:
    get:
      tags: [Files]
      security: [{ bearerAuth: [] }]
      summary: 工作空间文件变更（SSE）
      description: |
        `text/event-stream`；首包 `snapshot`，随后 `file_status` / `file_removed` / `processing_idle` / `error`。
        查询参数与 `GET /api/workspace` 相同。
      parameters:
        - name: folder
          in: query
          schema: { type: string }
        - name: folder_path
          in: query
          schema: { type: string }
        - name: type
          in: query
          description: semantic_type 过滤
          schema: { type: string }
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema: { type: string }
...